+++ package.json
@@ -39,11 +39,13 @@
     "react-native": "0.79.5",
     "react-native-calendars": "^1.1313.0",
     "react-native-gesture-handler": "~2.24.0",
+    "react-native-iap": "^13.0.4",
     "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-web": "~0.20.0",
-    "react-native-webview": "13.13.5"
+    "react-native-webview": "13.13.5",
+    "expo-device": "~7.1.4"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
